Improve stack allocation and disable most of low power mode code
1. Stack allocation now includes seemlessly the size required for the
context switch.

2. There was a bug in low power mode because it wasn't taking into
account that SR is restored from the stack when returning from
interrupt. The simplest way of fixing this is to use an idle thread.
We were trying to avoid using idle thread to save some bytes in memory,
but it would be on the expense of code readability and portability.
While a better solution is not found we are disabling LP code.

Signed-off-by: Alair Dias Junior <alair@alair.me>