Pack the HalfKOS runtime data into a structure
Instead of having the runtime data spread throughout the code
in several variables, we packed all data into a structure. There
is an inconvenient: acessing those data structures using basic
inline assembly is cumbersome. We are using extended ASM even
though GCC manual says it is not supported in naked functions.
We could have used basic assembly with fixed offsets to access
the structures, but this would be difficult at this time of the
project when data structures are changing a lot. We need to
revisit this decision in the future.

Signed-off-by: Alair Dias Junior <alair@alair.me>