Move entry point (main) inside of HalfKOS
This change moves the main function (entry point) inside of HalfKOS.
After initialization, HalfKOS will call a user setup function where
the user can create tasks and setup the system in general. This was
decided for allowing HalfKOS to configure the stack pointer before
the user starts calling functions, which may cause user's global
variables to be overwritten.

The other way around that issue would be defining a linker script
for HalfKOS but we are trying to avoid that to keep the system as
simple as possible for people to understand and port. Creating a
linker script is not something straight forward and the solution
with a setup function is very intuitive. If this decision creates
problems in the future, we will revisit it.

Signed-off-by: Alair Dias Junior <alair@alair.me>