Fix starting stack issue and improve memory organization
Before starting HalfKOS, stack is set according to the Scat file and
CRT. This may lead to global variables being overwritten by stack
(stack overflow). This change restarts the stack when HalfKOS is
initialized. Again, we could've done that on scat file, but we
prefer to do it explicitly on the code for better claritiy.

Signed-off-by: Alair Dias Junior <alair@alair.me>